# Speed-up: pass PJ_COORD by reference for fwd4d and inv4d callbacks