--- a/src/inv.cpp
+++ b/src/inv.cpp
@@ -217,7 +217,7 @@ PJ_LPZ pj_inv3d (PJ_XYZ xyz, PJ *P) {
 
 
 
-PJ_COORD pj_inv4d (PJ_COORD coo, PJ *P) {
+bool pj_inv4d (PJ_COORD& coo, PJ *P) {
 
     const int last_errno = P->ctx->last_errno;
     P->ctx->last_errno = 0;
@@ -225,7 +225,10 @@ PJ_COORD pj_inv4d (PJ_COORD coo, PJ *P) {
     if (!P->skip_inv_prepare)
         inv_prepare (P, coo);
     if (HUGE_VAL==coo.v[0])
-        return proj_coord_error ();
+    {
+        coo = proj_coord_error ();
+        return false;
+    }
 
     /* Call the highest dimensional converter available */
     if (P->inv4d)
@@ -242,13 +245,24 @@ PJ_COORD pj_inv4d (PJ_COORD coo, PJ *P) {
     }
     else {
         proj_errno_set (P, PROJ_ERR_OTHER_NO_INVERSE_OP);
-        return proj_coord_error ();
+        coo = proj_coord_error ();
+        return false;
     }
     if (HUGE_VAL==coo.v[0])
-        return proj_coord_error ();
+    {
+        coo = proj_coord_error ();
+        return false;
+    }
 
     if (!P->skip_inv_finalize)
         inv_finalize (P, coo);
 
-    return error_or_coord(P, coo, last_errno);
+    if (P->ctx->last_errno)
+    {
+        coo = proj_coord_error();
+        return false;
+    }
+
+    P->ctx->last_errno = last_errno;
+    return true;
 }