--- a/src/transformations/defmodel.cpp
+++ b/src/transformations/defmodel.cpp
@@ -343,34 +343,24 @@ static PJ *destructor(PJ *P, int errlev) {
     return pj_default_destructor(P, errlev);
 }
 
-static PJ_COORD forward_4d(PJ_COORD in, PJ *P) {
+static void forward_4d(PJ_COORD &coo, PJ *P) {
     auto *Q = (struct defmodelData *)P->opaque;
 
-    PJ_COORD out;
-    out.xyzt.t = in.xyzt.t;
-
-    if (!Q->evaluator->forward(Q->evaluatorIface, in.xyzt.x, in.xyzt.y,
-                               in.xyzt.z, in.xyzt.t, out.xyzt.x, out.xyzt.y,
-                               out.xyzt.z)) {
-        return proj_coord_error();
+    if (!Q->evaluator->forward(Q->evaluatorIface, coo.xyzt.x, coo.xyzt.y,
+                               coo.xyzt.z, coo.xyzt.t, coo.xyzt.x, coo.xyzt.y,
+                               coo.xyzt.z)) {
+        coo = proj_coord_error();
     }
-
-    return out;
 }
 
-static PJ_COORD reverse_4d(PJ_COORD in, PJ *P) {
+static void reverse_4d(PJ_COORD &coo, PJ *P) {
     auto *Q = (struct defmodelData *)P->opaque;
 
-    PJ_COORD out;
-    out.xyzt.t = in.xyzt.t;
-
-    if (!Q->evaluator->inverse(Q->evaluatorIface, in.xyzt.x, in.xyzt.y,
-                               in.xyzt.z, in.xyzt.t, out.xyzt.x, out.xyzt.y,
-                               out.xyzt.z)) {
-        return proj_coord_error();
+    if (!Q->evaluator->inverse(Q->evaluatorIface, coo.xyzt.x, coo.xyzt.y,
+                               coo.xyzt.z, coo.xyzt.t, coo.xyzt.x, coo.xyzt.y,
+                               coo.xyzt.z)) {
+        coo = proj_coord_error();
     }
-
-    return out;
 }
 
 // Function called by proj_assign_context() when a new context is assigned to