--- a/src/transformations/molodensky.cpp
+++ b/src/transformations/molodensky.cpp
@@ -260,10 +260,8 @@ static PJ_XYZ forward_3d(PJ_LPZ lpz, PJ *P) {
 }
 
 
-static PJ_COORD forward_4d(PJ_COORD obs, PJ *P) {
-    const auto xyz = forward_3d(obs.lpz, P);
-    obs.xyz = xyz;
-    return obs;
+static void forward_4d(PJ_COORD& obs, PJ *P) {
+    obs.xyz = forward_3d(obs.lpz, P);
 }
 
 
@@ -293,10 +291,8 @@ static PJ_LPZ reverse_3d(PJ_XYZ xyz, PJ *P) {
 }
 
 
-static PJ_COORD reverse_4d(PJ_COORD obs, PJ *P) {
-    const auto lpz = reverse_3d(obs.xyz, P);
-    obs.lpz = lpz;
-    return obs;
+static void reverse_4d(PJ_COORD& obs, PJ *P) {
+    obs.lpz = reverse_3d(obs.xyz, P);
 }